requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pypdf==5.1.0
openpyxl==3.1.5
openai==1.59.7
rq==2.1.0
redis==5.2.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
structlog==24.4.0
